pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "graphbot"
version = "0.1.0"
[tool.poetry.dependencies]
dnspython = "^1.16"
flask = "^1.1"
motor = "^2.1"
python = "^3.8"
schedule = "^0.6.0"
"discord.py" = "^1.4.1"
[tool.poetry.dev-dependencies]
